+++ examples/demo_theme_changer_all.py
@@ -1,0 +1,77 @@
+"""
+This is a minimal example of a theme switcher clientside callback with a toggle switch.
+Note - this is the version for dash-bootstrap-components V1.0.  and Dash V2.0
+
+The Bootstrap themed templates are from the dash-bootstrap-templates library.
+See more information at https://github.com/AnnMarieW/dash-bootstrap-templates
+    pip install dash-bootstrap-templates.
+"""
+
+from dash import Dash, dcc, html, Input, Output
+import pandas as pd
+import plotly.express as px
+import dash_bootstrap_components as dbc
+from aio import ThemeChangerAIO, template_from_url
+
+app = Dash(__name__, external_stylesheets=[dbc.themes.BOOTSTRAP, dbc.icons.BOOTSTRAP, dbc.icons.FONT_AWESOME ])
+
+
+df = pd.DataFrame(
+    {
+        "Fruit": ["Apples", "Oranges", "Bananas", "Apples", "Oranges", "Bananas"],
+        "Amount": [4, 1, 2, 2, 4, 5],
+        "City": ["SF", "SF", "SF", "Montreal", "Montreal", "Montreal"],
+    }
+)
+header = html.H4(
+    "ThemeSwitchAIO Demo", className="bg-primary text-white p-4 mb-2 text-center"
+)
+buttons = html.Div(
+    [
+        dbc.Button("Primary", color="primary", className="mr-1"),
+        dbc.Button("Secondary", color="secondary", className="mr-1"),
+        dbc.Button("Success", color="success", className="mr-1"),
+        dbc.Button("Warning", color="warning", className="mr-1"),
+        dbc.Button("Danger", color="danger", className="mr-1"),
+        dbc.Button("Info", color="info", className="mr-1"),
+        dbc.Button("Light", color="light", className="mr-1"),
+        dbc.Button("Dark", color="dark", className="mr-1"),
+        dbc.Button("Link", color="link"),
+    ],
+    className="m-4",
+)
+
+graph = html.Div(dcc.Graph(id="graph"), className="m-4")
+icon = html.Div(["FA", html.I(className = "bi bi-sun")])
+icon_fa = html.Div(["BS",html.I(className = "fa fa-sun")])
+
+"""
+===============================================================================
+Layout
+"""
+app.layout = dbc.Container(
+    [
+        header,
+        dbc.Row(
+            [
+                dbc.Col(ThemeChangerAIO(aio_id="theme", radio_props={"value":dbc.themes.VAPOR}), width=2,),
+                dbc.Col([buttons, graph, icon, icon_fa],),
+            ]
+        ),
+    ],
+    className="m-4",
+    fluid=True,
+)
+
+
+@app.callback(
+    Output("graph", "figure"), Input(ThemeChangerAIO.ids.radio("theme"), "value"),
+)
+def update_graph_theme(theme):
+    return px.bar(
+        df, x="Fruit", y="Amount", color="City", barmode="group", template=template_from_url(theme)
+    )
+
+
+if __name__ == "__main__":
+    app.run_server(debug=True)